Review response: Import/Export methods extracted into a mixin
