+++ app/models/generic_object_definition.rb
@@ -1,5 +1,6 @@
 class GenericObjectDefinition < ApplicationRecord
   include YAMLImportExportMixin
+  include_concern 'ImportExport'
 
   TYPE_MAP = {
     :boolean  => ActiveModel::Type::Boolean.new,
@@ -23,7 +24,6 @@ class GenericObjectDefinition < ApplicationRecord
   REG_ATTRIBUTE_NAME = /\A[a-z][a-zA-Z_0-9]*\z/
   REG_METHOD_NAME    = /\A[a-z][a-zA-Z_0-9]*[!?]?\z/
   ALLOWED_ASSOCIATION_TYPES = (MiqReport.reportable_models + %w(GenericObject)).freeze
-  IMPORT_CLASS_NAMES = %w[GenericObjectDefinition].freeze
 
   serialize :properties, Hash
 
@@ -152,53 +152,6 @@ class GenericObjectDefinition < ApplicationRecord
 
   def self.display_name(number = 1)
     n_('Generic Object Class', 'Generic Object Classes', number)
-  end
-
-  def export_to_array
-    god_attrs = attributes
-    ["id", "created_at", "updated_at"].each { |god| god_attrs.delete(god) }
-    [{self.class.to_s => god_attrs}]
-  end
-
-  def self.import_from_hash(god, options = nil)
-    raise _("No Generic Object Definition to Import") if god.nil?
-    if god["name"].blank? || god["properties"].blank?
-      raise _("Incorrect format.")
-    end
-    existing_god = GenericObjectDefinition.find_by(:name => god["name"])
-    if existing_god.present?
-      if options[:overwrite]
-        # if generic object definition exists, overwrite it's content
-        msg = "Overwriting Generic Object Definition: [#{existing_god.name}]"
-        existing_god.attributes = god
-        result = {
-          :message => "Replaced Generic Object Definition: [#{god["name"]}]",
-          :level   => :info,
-          :status  => :update
-        }
-      else
-        # if generic object definition exists dont overwrite
-        msg = "Skipping Generic Object Definition (already in DB): [#{existing_god.name}]"
-        result = {:message => msg, :level => :error, :status => :skip}
-      end
-    else
-      # create new generic object definition
-      msg = "Importing Generic Object Definition: [#{god["name"]}]"
-      existing_god = GenericObjectDefinition.new(god)
-      result = {
-        :message => "Imported Generic Object Definition: [#{god["name"]}]",
-        :level   => :info,
-        :status  => :add
-      }
-    end
-    _log.info(msg)
-
-    if result[:status].in?([:add, :update])
-      existing_god.save!
-      _log.info("- Completed.")
-    end
-
-    return god, result
   end
 
   private
